extract classes out of package.scala
