+++ tlair/src/main/scala/at/forsyte/apalache/tla/lir/TlaValue.scala
@@ -1,0 +1,7 @@
+package at.forsyte.apalache.tla.lir
+
+/**
+ * The base class for the universe of values (integers, Booleans, strings) used in TLA+.
+ * See the implementations in at.forsyte.apalache.tla.lir.values.
+ */
+abstract class TlaValue
